+++ zellij-utils/assets/prost/api.event.rs
@@ -666,6 +666,8 @@ pub struct TabInfo {
     pub has_bell_notification: bool,
     #[prost(bool, tag="19")]
     pub is_flashing_bell: bool,
+    #[prost(message, optional, tag="20")]
+    pub ui_theme: ::core::option::Option<super::style::Styling>,
 }
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]