+++ zellij-utils/src/errors.rs
@@ -403,6 +403,7 @@ pub enum ScreenContext {
     WriteToPaneId,
     Paste,
     SetPaneColor,
+    SetUiTheme,
     WriteKeyToPaneId,
     CopyTextToClipboard,
     MovePaneWithPaneId,